libpcp_web: code cleanup for the non-libuv build case
