+++ src/libpcp_web/src/nodiscover.c
@@ -14,6 +14,7 @@
 
 /* Stub routines for platforms lacking necessary features (UV, SSL) */
 #include "pmwebapi.h"
+#include "discover.h"
 
 int
 pmDiscoverRegister(const char *path, pmDiscoverModule *module,
@@ -27,4 +28,58 @@ void
 pmDiscoverUnregister(int handle)
 {
     (void)handle;
+}
+
+discoverModuleData *
+getDiscoverModuleData(pmDiscoverModule *module)
+{
+    (void)module;
+    return NULL;
+}
+
+int
+pmDiscoverSetSlots(pmDiscoverModule *module, void *slots)
+{
+    (void)slots; (void)module;
+    return -EOPNOTSUPP;
+}
+
+int
+pmDiscoverSetConfiguration(pmDiscoverModule *module, dict *config)
+{
+    (void)module; (void)config;
+    return -EOPNOTSUPP;
+}
+
+int
+pmDiscoverSetEventLoop(pmDiscoverModule *module, void *events)
+{
+    (void)module; (void)events;
+    return -EOPNOTSUPP;
+}
+
+void
+pmDiscoverSetupMetrics(pmDiscoverModule *module)
+{
+    (void)module;
+}
+
+int
+pmDiscoverSetMetricRegistry(pmDiscoverModule *module, mmv_registry_t *registry)
+{
+    (void)module; (void)registry;
+    return -EOPNOTSUPP;
+}
+
+int
+pmDiscoverSetup(pmDiscoverModule *module, pmDiscoverCallBacks *cbs, void *arg)
+{
+    (void)module; (void)cbs; (void)arg;
+    return -EOPNOTSUPP;
+}
+
+void
+pmDiscoverClose(pmDiscoverModule *module)
+{
+    (void)module;
 }
